add offload_folder

+++ moondream/text_model.py
@@ -33,7 +33,7 @@ class TextModel:
         embeds = []
 
         def _add_toks(toks):
-            embeds.append(self.text_emb(toks))
+            embeds.append(self.text_emb(toks.float()))
 
         def _tokenize(txt):
             return self.tokenizer(
@@ -46,13 +46,13 @@ class TextModel:
         )
 
         if "<image>" not in prompt:
-            embeds.append(self.text_emb(_tokenize(prompt)))
+            embeds.append(self.text_emb(_tokenize(prompt).float()))
         else:
             assert prompt.count("<image>") == 1
             before, after = prompt.split("<image>")
-            embeds.append(self.text_emb(_tokenize(f"{before}<image>")))
+            embeds.append(self.text_emb(_tokenize(f"{before}<image>").float())
             embeds.append(image_embeds.to(self.model.device))
-            embeds.append(self.text_emb(_tokenize(f"</image>{after}")))
+            embeds.append(self.text_emb(_tokenize(f"</image>{after}").float())
 
         return torch.cat(embeds, dim=1)
